Added in the new backend as go_backend
I did not include the db or the .env files so auth is currently not
working. I figured it would be easier to generate new files per person.

### For the newly kind of formed fullstack team

Templating is in github.com/a-h/templ. echo is a sumpler way of working
with go requests but to be honest we may remove this dependency down the
line because ist not super necessary. Lastly I use a little bit of
alpine for writing javascript which is a tiny libary similar to
JQuery... I just don't like javascript so don't feel its necessary to
use it. Hypermedia was the inspuration behind this code. I wrote little
javascript since we are not netflix :)