Added in user id to the databse table and including database now. Moved update user
Realised why most database have an id now. When the user is logged in
and goes to the settings form the form url for updating the details is
found at PATCH /settings/some-uuid-that-is-random. The reason for this
is we don't need to get the user details from a token. The state is the
entity id and the action we can take on it is patching

I think there must be a way to include the values in the form as a
default when you navigate but not sure yet.

Lastly there is a database included in the backend sqlite if you want to
use it for testing crud operations