Fixed issues with map loads.
Rewrote map initialisation so is is iniliatised on google maps script
download. The map initialisation now checks for the presence of a map
tag before writing the map to the dom.

If for example you are on the dashboard and navigate to the map this
creates an issue where you need to run the initialise map again since
the dom element is now available. that is where HX-Trigger-After-Swap
comes in. This is a header which triggers a dispatch of an event names
"mapUpdate" and then on the body tag I listen for this update event and
run the map initialisation again.