Added filtering for real_estate_price_data
For pull request #37  on issue #33

Added handler processing for the real_estate_price_data file.

PARSING IS NOT WORKING RIGHT NOW -> Json data for coordinates is a string, so I will have to transform it in the backend or ask the Data team to change it to float.

http filter options can be queried/combined for the following:

- zipcode
- homevalue
- income
- age