Auth works, pleaase see details
After evaluating the implementation of a single page application with an
API and our own database, we found that cookie-based authentication is
not suitable for this architecture. Cookie authentication works best in
a monolithic application and doesn't integrate well with a separate Go
API and SPA. Therefore, we have decided to revert back to JSON Web Token
for authentication. At this stage of the project, ensuring everything
functions correctly is more important than creating a custom backend
solution.